Only use unsigned int for the opengl array, never use unsigned short
This simplify the code, I don't think it introduce a huge speed penalty.

+++ src/modules/Atmosphere.hpp
@@ -86,7 +86,7 @@ private:
 
 	Vec2f* posGrid;
 	Vec4f* colorGrid;
-	GLushort* indices;
+	GLuint* indices;
 
 	//! The average luminance of the atmosphere in cd/m2
 	float averageLuminance;
